requirements.txt
scapy>=2.5.0
pandas>=1.5.0
numpy>=1.22.0
requests>=2.28.0
PyYAML>=6.0
jinja2>=3.1.0
geoip2>=4.0.0
matplotlib>=3.6.0
scikit-learn>=1.1.0
pytest>=7.0.0
python-dotenv>=0.21.0